+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Protocol: gossipd now deliberately delays spamming with `channel_update`.
 - Config: `--conf` option to set config file.
 - JSON API: Added description to invoices and payments (#1740).
+- pylightning: RpcError now has `method` and `payload` fields.
 
 ### Changed